Add help command
There is no way for new users to learn how to use the application.

Makes it difficult for new users to adopt the application.

Let's add a helpsheet to let users know what commands they can use, and
how to use these commands, by adding a method in Ui to print it.

We handle the printing in Ui as we do not have a separate class for
commands specifically, so we print it from the Ui class to ensure that
we only have to change that one method when we if we want to make any
changes to the helpsheet